scripts/requirements.txt
pwntools==4.4.0
pwn==1.0
pycryptodome==3.10.1
